3. Доменная модель проекта "АвтоМаг" [#504897 # [#504897]]

+++ db/changelog/scripts/script1_auto_user.sql
@@ -1,5 +1,5 @@
 CREATE TABLE auto_user(
-    id int primary key auto_increment,
+    id bigint primary key auto_increment,
     login varchar(100),
     password varchar(100)
 );
